SecSphr--new-main/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.10.0
Werkzeug==3.1.3